Refactor x86 initialization code quite a lot.
