+++ kernel/src/main.rs
@@ -50,10 +50,10 @@ pub mod arch;
 #[path = "arch/x86_64/mod.rs"]
 pub mod x86_64_arch;
 
+mod environment;
 mod error;
 mod fs;
 mod graphviz;
-mod kcb;
 mod memory;
 mod nr;
 mod nrproc;
